- multikin checkboxes for B-factor and occupancy were separated for JSR - various cosmetic fixes for JSR - fixed JS model selection form on welcome.php to work in Firefox - modified session expiration date tracking in lib/sessions.php   Added MP_SESSION_LIFETIME_EXT to config; use it in pages/save_session   Fixed admin/show_sessions to use the new last-touched vs. TTL data - changed setup.sh to avoid clobbering config.php
git-svn-id: file:///home/smlewis/svn/git-mp3/another_svn/tmp/tmp/trunk/molprobity3@158 5035da28-9fce-0310-abbb-aec50abfc0e8

+++ setup.sh
@@ -18,7 +18,11 @@ if [ ! -f setup.sh ]; then
 fi
 
 # Create config file
-cp config/config.php.sample config/config.php
+if [ ! -f config/config.php ]; then
+    cp config/config.php.sample config/config.php
+else
+    echo "config.php already exists; no changes made"
+fi
 
 # Set world-writable permisions on data/ and tmp/
 chmod 777 public_html/data
